Provide option to not panic on reconnect retry
Adds a `Reconnect()` method to force a re-connect.
This can be used once max retries have been hit on attempting to
reconnect to an endpoint.

The caller knows to call `Reconnect()` because posting data will
generate a `ConnectionError` error in the case where the connection is
down.

`Reconnect()` blocks until either max reconnect attemps have been
reached again or a connection is successful.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>