Test how the logger behaves in async mode
New testcases have been written to test if PostWithTime() and Close()
are working as expected. More specifically, they test if Close() succeed
with all combination of options (Async + ForceStopAsyncSend).

Tests on PostWithTime() are divided into two parts:

* With transient network failure ;
* And without any ;

And tests on Close() are divided into two parts:

* CloseOnAsyncConnect: When no logs have been written yet ;
* CloseOnAsyncReconnect: When there're pending logs to write ;

Finally, the test `Test_send_WritePendingToConn` has been removed
because it was relying too much on the logger internals to test the
logger's behavior from an API user PoV. And the init() that was starting
a tcp listener has been removed to make the tests easier to execute
(otherwise, we've to wait for go test to timeout before re-running it).

Signed-off-by: Albin Kerouanton <albin.kerouanton@knplabs.com>