Fix panic on writing to closed channel
The docker daemon RingLogger implementation is written in such a way
that it can call Close() on the underlying driver, without stopping to
write to it. This can cause panics on fluent-golang-logger which has no
checks on whether the channel was closed by the daemon. The contract on
the logger.Logger interface in docker doesn't specify that it will make
sure that the call to Close() and Log() will not compete so other
logging driver implementations have similar semantics as this commit to
account for that.

Signed-off-by: Anuj Varma <anujvarma@thumbtack.com>