Do not allow writing events after fluent.Close()
This implements a proposal by Albin Kerouanton.

The behavior of `Fluent.PostWithTime()` is a bit odd in sync mode;
It will silently reopen connections, even if the client has been
finalized via `Fluent.Close()`.

Change the behavior so that `PostWithTime()` returns an error
instead of doing an implicit reconnection in such a case.

Signed-off-by: Fujimoto Seiji <fujimoto@ceptord.net>