fluent: pending channel thread saftey test
Added a new test to ensure that pending channel is accessed in a
thread safe manner. The code can be simplified by removing the
pendingMutex lock. There's a risk in this codebase wrt how muconn
and pendingMutex are acquired and released in different orders in
different methods.

This test is a precursor to the change to remove the pendingMutex
to ensure that nothing is broken.

Signed-off-by: Anirudh Aithal <aithal@amazon.com>